+++ src/main/java/comparator/CompTeamOPPG.java
@@ -11,6 +11,6 @@ import simulation.Team;
 public class CompTeamOPPG implements Comparator<Team> {
     @Override
     public int compare(Team a, Team b) {
-        return a.teamOppPoints / a.numGames() < b.teamOppPoints / b.numGames() ? -1 : a.teamOppPoints / a.numGames() == b.teamOppPoints / b.numGames() ? 0 : 1;
+        return (float) a.teamOppPoints / a.numGames() < (float) b.teamOppPoints / b.numGames() ? -1 : (float) a.teamOppPoints / a.numGames() == (float) b.teamOppPoints / b.numGames() ? 0 : 1;
     }
 }
